move mime classifier unit tests into unit test crate
