+++ tests/unit/net/mod.rs
@@ -5,4 +5,5 @@
 mod cookie;
 mod data_loader;
 mod image_cache_task;
+mod mime_classifier;
 mod resource_task;
